+++ R/get_channel_stats.R
@@ -4,9 +4,9 @@
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
 #' @return nested named list with top element names:
-#' kind, etag, id, snippet (list of details of the channel including title), statistics (list of 5)
+#' \code{kind, etag, id, snippet (list of details of the channel including title), statistics (list of 5)}
 #' 
-#' If the channel_id is mistyped or there is no information, an empty list is returned
+#' If the \code{channel_id} is mistyped or there is no information, an empty list is returned
 #' 
 #' @export
 #' 
@@ -17,12 +17,13 @@
 #' 
 #' # Set API token via yt_oauth() first
 #' 
-#' get_channel_stats(channel_id="UChTJTbr5kf3hYazJZO-euHg")
+#' get_channel_stats(channel_id="UCMtFAi84ehTSYSE9XoHefig")
+#' get_channel_stats(channel_id="UCMtFAi84ehTSYSE9Xo") # Incorrect channel ID
 #' }
 
 get_channel_stats <- function (channel_id = NULL, ...) {
 
-	if (is.null(channel_id)) stop("Must specify a channel ID")
+	if (!is.character(channel_id)) stop("Must specify a channel ID.")
 	
 	querylist <- list(part = "statistics,snippet", id = channel_id)