+++ R/get_comments.R
@@ -7,11 +7,11 @@
 #' \code{parent_id}: parent ID.
 #'  
 #' @param part  Comment resource requested. Required. Comma separated list of one or more of the 
-#' following: \code{id, snippet}. e.g., "id, snippet", "id" Default: \code{snippet}.  
+#' following: \code{id, snippet}. e.g., \code{"id, snippet"}, \code{"id"}, etc. Default: \code{snippet}.  
 #' @param simplify Data Type: Boolean. Default is TRUE. If TRUE, the function returns a data frame. Else a list with all the information returned.
 #' @param max_results  Maximum number of items that should be returned. Integer. Optional. Can be between 20 and 100. Default is 100.
 #' @param page_token  Specific page in the result set that should be returned. Optional.
-#' @param text_format Data Type: Character. Default is "html". Only takes "html" or "plainText." Optional. 
+#' @param text_format Data Type: Character. Default is \code{"html"}. Only takes \code{"html"} or \code{"plainText"}. Optional. 
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #'  
 #' @return 
@@ -19,9 +19,10 @@
 #' Within each of the \code{items} is an item \code{snippet} which has an item \code{topLevelComment$snippet$textDisplay}
 #' that contains the actual comment.
 #' 
-#' When filter is comment_id, and simplify is TRUE, and there is a correct comment id, it returns a data.frame with 
-#' following cols:  id, authorDisplayName, authorProfileImageUrl, authorChannelUrl, value, textDisplay, canRate, viewerRating, likeCount
-#' publishedAt, updatedAt
+#' When filter is \code{comment_id}, and \code{simplify} is \code{TRUE}, and there is a correct comment id, it returns a \code{data.frame} with 
+#' following cols: 
+#' \code{id, authorDisplayName, authorProfileImageUrl, authorChannelUrl, value, textDisplay, canRate, viewerRating, likeCount
+#' publishedAt, updatedAt}
 #'  
 #' @export
 #' @references \url{https://developers.google.com/youtube/v3/docs/comments/list}
@@ -33,12 +34,13 @@
 #' 
 #' get_comments(filter = c(comment_id="z13dh13j5rr0wbmzq04cifrhtuypwl4hsdk"))
 #' get_comments(filter = 
-#' c(comment_id="z13dh13j5rr0wbmzq04cifrhtuypwl4hsdk, z13dh13j5rr0wbmzq04cifrhtuypwl4hsdk"))
+#' c(comment_id="z13dh13j5rr0wbmzq04cifrhtuypwl4hsdk, 
+#' 		         z13dh13j5rr0wbmzq04cifrhtuypwl4hsdk"))
 #' }
 
-get_comments <- function (filter = NULL, part = "snippet", text_format = "html", simplify = TRUE, max_results = 100, page_token = NULL, ...) {
+get_comments <- function (filter = NULL, part = "snippet", max_results = 100, text_format = "html", page_token = NULL, simplify = TRUE, ...) {
 
-	if (max_results < 20 | max_results > 100) stop("max_results only takes a value between 20 and 100")
+	if (max_results < 20 | max_results > 100) stop("max_results only takes a value between 20 and 100.")
 	if (text_format != "html" & text_format !="plainText") stop("Provide a legitimate value of textFormat.")
 
 	if (!(names(filter) %in% c("parent_id", "comment_id"))) stop("filter can only take one of values: comment_id, parent_id.")
@@ -59,11 +61,9 @@ get_comments <- function (filter = NULL, part = "snippet", text_format = "html",
     	return(list())
     }
 
-	res     <- raw_res$items[[1]]$snippet
-
 	if (simplify==TRUE & part=="snippet") {
 		simple_res  <- lapply(raw_res$items, function(x) unlist(x$snippet))
-		simpler_res <- as.data.frame(do.call(rbind, simple_res))
+		simpler_res <- ldply(simple_res, rbind)
 		simpler_res$id <- raw_res$items[[1]]$id
 		return(simpler_res)
 	}