+++ R/get_stats.R
@@ -3,7 +3,7 @@
 #' @param video_id Character. Id of the video. Required.
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
-#' @return list with 6 elements: id, viewCount, likeCount, dislikeCount, favoriteCount, commentCount
+#' @return list with 6 elements: \code{id, viewCount, likeCount, dislikeCount, favoriteCount, commentCount}
 #'
 #' @export
 #' 
@@ -19,7 +19,7 @@
 
 get_stats <- function (video_id=NULL, ...) {
 
-	if (is.null(video_id)) stop("Must specify a video ID")
+	if (!is.character(video_id)) stop("Must specify a video ID.")
 	
 	querylist <- list(part="statistics", id = video_id)
     
@@ -39,5 +39,5 @@ get_stats <- function (video_id=NULL, ...) {
 	cat('No. of Favorites', stat_res$favoriteCount, "\n")
 	cat('No. of Comments', stat_res$commentCount, "\n")
  
-	return(invisible(c(id=res$id, stat_res)))
+	c(id=res$id, stat_res)
 }