+++ R/get_video_details.R
@@ -5,8 +5,8 @@
 #' @param video_id Comma separated list of IDs of the videos for which details are requested. Required. 
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
-#' @return list with the following elements: id (video id that was passed), publishedAt, channelId, title, description, thumbnails, 
-#' channelTitle, categoryId, liveBroadcastContent, localized, defaultAudioLanguage
+#' @return list with the following elements: \code{id} (video id that was passed), \code{publishedAt, channelId, title, description, thumbnails, 
+#' channelTitle, categoryId, liveBroadcastContent, localized, defaultAudioLanguage}
 #' 
 #' @export
 
@@ -22,7 +22,7 @@
 
 get_video_details <- function (video_id = NULL, ...){
 	
-	if (is.null(video_id)) stop("Must specify a video ID")
+	if (!is.character(video_id)) stop("Must specify a video ID.")
 
 	querylist <- list(part="snippet", id = video_id)