+++ R/list_abuse_report_reasons.R
@@ -1,16 +1,16 @@
 #' List reasons that can be used to report abusive videos
 #' 
 #' @param part  Caption resource requested. Required. Comma separated list of one or more of the 
-#' following: \code{id, snippet}. e.g., "id, snippet", "id" Default: \code{snippet}.  
+#' following: \code{id, snippet}. e.g., \code{"id, snippet"}, \code{"id"}, etc. Default: \code{snippet}.  
 #' @param hl  Language used for text values. Optional. Default is \code{en-US}. For other allowed language codes, see \code{\link{list_langs}}.
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #'  
 #' @return 
 #' 
 #' If no results, empty data.frame returned
-#' If part requested = "id, snippet", data.frame with 4 columns: etag, id, label, secReasons
-#' If part requested = "snippet", data.frame with 3 columns: etag, label, secReasons
-#' If part requested = "id", data.frame with 2 columns: etag, id
+#' If part requested = "id, snippet" or "snippet", 
+#' \code{data.frame} with 4 columns: \code{etag, id, label, secReasons}
+#' If part requested = "id", data.frame with 2 columns: \code{etag, id}
 #' 
 #' @export
 #' @references \url{https://developers.google.com/youtube/v3/docs/videoAbuseReportReasons/list}
@@ -26,27 +26,24 @@
 
 list_abuse_report_reasons <- function (part = "id, snippet", hl = "en-US", ...) {
 
+	if (!is.character(part)) stop("Provide a valid value for part.")
+
 	querylist <- list(part=part)
 
 	res <- tuber_GET("videoAbuseReportReasons", querylist, ...)
 
-	resdf <- NULL
+	resdf <- data.frame()
 
 	if (length(res$items) != 0) {
 		
-		if (part=="id, snippet") {
+		if (part=="id, snippet" | part=="snippet") {
 			simple_res  <- lapply(res$items, function(x) c(etag=x$etag, id=x$id, label=x$snippet$label, secReasons=paste(unlist(x$snippet$secondaryReasons), collapse=",")))
-			resdf       <- as.data.frame(do.call(rbind, simple_res))
-		}
-
-		if (part=="id, snippet") {
-			simple_res  <- lapply(res$items, function(x) c(etag=x$etag, label=x$snippet$label, secReasons=paste(unlist(x$snippet$secondaryReasons), collapse=",")))
-			resdf       <- as.data.frame(do.call(rbind, simple_res))
+			resdf       <- ldply(simple_res, rbind)
 		}
 
 		if (part=="id") {
 			simple_res  <- lapply(res$items, function(x) c(etag=x$etag, id=x$id))
-			resdf       <- as.data.frame(do.call(rbind, simple_res))
+			resdf       <- ldply(simple_res, rbind)
 		}
 
 	}