+++ R/list_caption_tracks.R
@@ -8,9 +8,9 @@
 #' @param simplify Boolean. Default is TRUE. When TRUE, and part is \code{snippet}, a data.frame is returned
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
-#' @return list of caption tracks. When simplify is TRUE, a data.frame is returned with 
-#' following columns: videoId, lastUpdated, trackKind, language, name, audioTrackType, isCC,
-#' isLarge, isEasyReader, isDraft, isAutoSynced, status, id (caption id)
+#' @return list of caption tracks. When \code{simplify} is \code{TRUE}, a \code{data.frame} is returned with 
+#' following columns: \code{videoId, lastUpdated, trackKind, language, name, audioTrackType, isCC,
+#' isLarge, isEasyReader, isDraft, isAutoSynced, status, id} (caption id)
 #' 
 #' @export
 #' 
@@ -26,7 +26,7 @@
 
 list_caption_tracks <- function (part="snippet", video_id=NULL, lang="en", id = NULL, simplify = TRUE, ...) {
 
-	if (is.null(video_id)) stop("Must specify a video ID")
+	if (!is.character(video_id)) stop("Must specify a video ID.")
 	
 	querylist = list(part=part, videoId = video_id, id = id)
 	raw_res <- tuber_GET("captions", query = querylist, ...)
@@ -37,7 +37,7 @@ list_caption_tracks <- function (part="snippet", video_id=NULL, lang="en", id = 
     }
 
     if (simplify == TRUE & part=="snippet") {
-    	res_df 	  <- as.data.frame(do.call(rbind, lapply(raw_res$items, function(x) unlist(x$snippet))))
+    	res_df 	  <- ldply(lapply(raw_res$items, function(x) unlist(x$snippet)), rbind)
     	res_df$id <- sapply(raw_res$items, function(x) unlist(x$id))
     	return(res_df)
     }