+++ R/list_langs.R
@@ -3,7 +3,7 @@
 #' @param hl  Language used for text values. Optional. Default is \code{en-US}. For other allowed language codes, see \code{\link{list_langs}}.
 #' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
 #' 
-#' @return data.frame with 3 columns: hl (two letter abbreviation), name (of the language), etag
+#' @return data.frame with 3 columns: \code{hl} (two letter abbreviation), \code{name} (of the language), \code{etag}
 #' 
 #' @export
 #' 
@@ -17,23 +17,21 @@
 #' list_langs()
 #' }
 
-list_langs <- function (hl=NULL, ...) {
+list_langs <- function (hl = NULL, ...) {
 
-	querylist <- list(part="snippet", hl = hl)
+	querylist <- list(part = "snippet", hl = hl)
 
 	res <- tuber_GET("i18nLanguages", querylist)
 
-	resdf <- NA
+	resdf <- data.frame()
 
 	if (length(res$items) != 0) {
 		simple_res  <- lapply(res$items, function(x) c(unlist(x$snippet), etag=x$etag))
 		resdf       <- as.data.frame(do.call(rbind, simple_res))
-	} else {
-		resdf <- 0
-	}
+	} 
 
 	# Cat total results
 	cat("Total Number of Languages:", length(res$items), "\n")
 
-	return(invisible(resdf))
+	resdf
 }