+++ R/list_regions.R
@@ -1,12 +1,16 @@
 #' List Content Regions That Youtube Currently Supports
 #' 
-#' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
-#' @param hl  Language used for text values. Optional. Default is \code{en-US}. For other allowed language codes, see \code{\link{list_langs}}.
 #' 
-#' @return data.frame with 3 columns: gl (two letter abbreviation), name (of the region), etag
+#' @param hl  Language used for text values. Optional. Default is \code{en-US}. For other allowed language codes, see \code{\link{list_langs}}.
+#' @param \dots Additional arguments passed to \code{\link{tuber_GET}}.
+#' 
+#' @return data.frame with 3 columns: 
+#' \code{gl} (two letter abbreviation), \code{name} (of the region), \code{etag}
 #' 
 #' @export
+#' 
 #' @references \url{https://developers.google.com/youtube/v3/docs/i18nRegions/list}
+#'
 #' @examples
 #' \dontrun{
 #' 
@@ -17,21 +21,19 @@
 
 list_regions <- function (hl = NULL, ...) {
 
-	querylist <- list(part="snippet")
+	querylist <- list(part = "snippet")
 
 	res <- tuber_GET("i18nRegions", querylist, ...)
 
-	resdf <- NA
-
-	if (length(res$items) != 0) {
-		simple_res  <- lapply(res$items, function(x) c(unlist(x$snippet), etag=x$etag))
-		resdf       <- as.data.frame(do.call(rbind, simple_res))
-	} else {
-		resdf <- 0
-	}
+	resdf <- read.table(text = "", col.names = c("gl", "name", "etag"))
 
 	# Cat total results
 	cat("Total Number of Content Regions:", length(res$items), "\n")
 
-	return(invisible(resdf))
+	if (length(res$items) != 0) {
+		simple_res  <- lapply(res$items, function(x) c(unlist(x$snippet), etag=x$etag))
+		resdf       <- ldply(simple_res, rbind)
+	}
+
+	resdf
 }